Updated AppVeyor testing
Refactored AppVeyor testing to mirror that of the TravisCI testing. The
new style is much more manageable and clearly shows the future
potential for more aggressive shell testing (once all of the Windows
kinks are resolved).

There are still certain obvious issues or possible major issues in
relation to the how the conda build deals with the shell-agnostic-ness
of this PR.