Fix issue with creating new envs
Before the command `conda create -n testenv python` was returning an
empty list of packages to install in the new environment. This was
caused by filtering out the python package since it was already
installed in the root env. This behaviour has been updated such that the
filtering no longer takes place. So, the filtering behaviour will have
to be reimplemented, _probably_ after conda reaches the step where it
determines all envs