utils: Add sys_prefix_unfollowed
.. and use it in find_commands.py to account for the fact that conda
is installed into sub-environments as a symlink only and that sys.prefix
follows those symlinks meaning it does not give the environment's prefix

There may be other parts of the codebase that need to use this function
but find_commands.py fails to find external programs without this. It is
necessary to search both sys_prefix_unfollowed *and* sys.prefix in this
case as conda-build will often not be installed into the environment.

The failure here was seen when trying to run the test for conda-build-all
while building it through conda-build.

Do we really need to add C:\cygwin\bin here too? I hope not, at least not
in the non-'backwards compatibility' function.