Allow to link packages from extracted tarballs if tarball is gone
Currently, when linking a package, an already-extracted package is only
used when the corresponding tarball also exists and its md5 checksum
matches what is in the repodata. This means that if the tarball is not
present and only the extracted directory exists, conda always re-downloads
the tarball (and re-extracts it) even if the extracted directory exists
already because it can't compare the checksums otherwise.

This used to be different in previous versions such as 4.2.x where conda
would link a package from an extracted directory even if the tarball was
gone.

The new behavior is somewhat unfortunate in the sense that after a Miniconda
install, the pkgs directory is left with only extracted directories and no
tarballs. Also, `conda clean --tarballs` removes *all* tarballs and so causes
a re-download of all packages next time an environment is being created.

This PR modifies conda's behavior leveraging the newly introduced (in #5148)
`repodata_record.json` file which is written at package extraction time.
Among other things, it contains the md5 checksum of the tarball that was
extracted, and therefore allows us to perform the md5 comparison even in
the absence of the actual tarball.