Don't look into same cache when finding match for urls.txt
When making a urls.txt entry for an explicitly installed package,
conda searches its package cache for a similar package so that it can
take the origin information from that cache entry, assuming it is more
accurate than the path of the explicitly provided tarball path. However,
conda did not use to look into the cache that it was installing the
explicitly specified package into because before it starts looking for
a match it moves a possibly conflicting package out of the way.

However, since now we are also taking checksums from extracted
directories, conda found the extracted directory and matched it with the
explicit package to be installed. This caused installation of the
explicitly specified package to fail.