Preserve status in fish shell
The fish_prompt function typically does things like changing the color
of the prompt based on the exit status of the last command. This
status is however lost in the function defined by conda, since the
status of the last operation is always zero.

This fixes the issue by defining an identity function that returns the
last status (thereby setting the $status variable) before the original
fish_prompt function is called.