Remove reference to 'system' channel.
This channel no longer exists, and looks like an artifact left over in
the doc from a conda 2.x feature that is no longer supported nor
necessary (.condarc specified channels are search by default). This came
up on the conda mailing list as a point of confusion, and is confusing
since the term 'system' is already overloaded with multiple meanings.